fix(ci): Moved the helpers for getting the latest package version and the submodule and function name to a separate file as they're required by both run_tests and label_failures
